% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGE_result.R
\docType{data}
\name{DGE_result}
\alias{DGE_result}
\title{DGE results from psoriasis dataset.}
\format{
A data frame with 682 rows and 6 variables:
\describe{
\item{Gene}{The name of genes.}
\item{baseMean}{Mean of normalized counts for all samples}
\item{log2FoldChange}{Log2 fold change.}
\item{lfcSE}{Standard error of log2 fold change.}
\item{pvalue}{Wald test p-value.}
\item{padj}{BH adjusted p-values.}
}
}
\usage{
DGE_result
}
\description{
The DGE analysis results on 682 genes in 23010 T-cells of 3 psoriasis patients. The original records of cells are derived from the single-cell RNA-seq dataset collected by Reynolds et al. (2021) in the study of inflammatory skin diseases.
}
\examples{
data(DGE_result)
head(DGE_result)
}
\keyword{datasets}
