% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ghost_Basil.R
\name{Ghost_Basil}
\alias{Ghost_Basil}
\title{GhostBASIL solver}
\usage{
Ghost_Basil(b, Sigma_S, S, M, lambda = NULL)
}
\arguments{
\item{b}{Coefficient of linear term.}

\item{Sigma_S}{Correlation matrix \eqn{Sigma} minus the diagonal matrix \eqn{S}}

\item{S}{Diagnoal elements of the S-matrix.}

\item{M}{The number of knockoff copies (under the multiple knockoff framework).}

\item{lambda}{User-sepcific tuning parameter list.}
}
\value{
A list containing all considered tuning parameters \eqn{\lambda}'s and the corresponing solutions.
}
\description{
This function fits a lasso regression model between disease of interest and gene expression levels (observed and knockoff copies). It solves the quadratic optimization problem
\deqn{\min_{\beta}\frac{1}{2}\cdot\beta^TG\beta-\beta^Tb+\lambda\|\beta\|_1,}
over a set of tuning parameters \eqn{\lambda}'s, where
\eqn{G=(1_{M+1}1_{M+1}^T)\otimes\Sigma_{S}+I_{M+1}\otimes S}.
}
