% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_posdef.R
\name{is_posdef}
\alias{is_posdef}
\title{Positive semidefiniteness detection}
\usage{
is_posdef(A)
}
\arguments{
\item{A}{The input matrix}
}
\value{
An indicator whether the input matrixis positive semidefinite or not.
}
\description{
This function detects whether a input matrix is positive semidefinite or not
}
