% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnockoffCDGE.R
\name{KnockoffCDGE}
\alias{KnockoffCDGE}
\title{Knockoff-based conditional differential gene expression (CDGE) analysis}
\usage{
KnockoffCDGE(
  Z,
  Sigma,
  M,
  n,
  method = c("ME", "SDP", "SDP_no_julia", "PCA"),
  Gene_info = NULL,
  verbose = TRUE,
  tol = 0.001
)
}
\arguments{
\item{Z}{DGE Z-scores.}

\item{Sigma}{Correlation matrix.}

\item{M}{The number of knockoff copies (under the multiple knockoff framework).}

\item{n}{Sample size.}

\item{method}{The knockoff method used.}

\item{Gene_info}{Gene information.}

\item{verbose}{Indicator of whether progress of computation is plotted.}

\item{tol}{Convergence tolerance}
}
\value{
Knockoff q-values of all features (the smallest target FDR level for each feature to be selected).
}
\description{
This function performs the knockoff-based CDGE analysis with Z-scores of DGE analysis and correlation matrix of gene expressions.
}
\examples{
data("DGE_result")
data("Sigma")
set.seed(433)
Z<-Z_calculation(DGE_result$pvalue,DGE_result$log2FoldChange)
KnockoffCDGE(Z,Sigma,M=5,n=23010,method="ME",Gene_info=DGE_result)
}
