% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KO_filter.R
\name{KO_Filter}
\alias{KO_Filter}
\title{Knockoff q-values computation}
\usage{
KO_Filter(tau, kappa, M)
}
\arguments{
\item{tau}{Feature statistics (magnitude)}

\item{kappa}{Feature statistics (sign)}

\item{M}{The number of knockoff copies (under the multiple knockoff framework)}
}
\value{
Knockoff q-values of all features (the smallest target FDR level for each feature to be selected).
}
\description{
This function computes knockoff q-values from feature statistics of all features.
}
\examples{
set.seed(433)
tau<-c(runif(20)+0.5,runif(80))
kappa<-c(sample(0:5,20,prob=c(0.8,rep(0.04,5)),replace=TRUE),sample(0:5,80,replace=TRUE))
KO_Filter(tau,kappa,M=5)
}
