% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_calculation_ME.R
\name{S_calculation_ME}
\alias{S_calculation_ME}
\title{S-matrix computation (maximizing the entropy)}
\usage{
S_calculation_ME(Sigma, M, verbose = TRUE, tol = 0.001)
}
\arguments{
\item{Sigma}{Correlation matrix.}

\item{M}{The number of knockoff copies (under the multiple knockoff framework).}

\item{verbose}{Indicator of whether progress of computation is plotted.}

\item{tol}{Convergence tolerance}
}
\value{
Diagnoal elements of the S-matrix for knockoff generation.
}
\description{
This function computes the S-matrix from correlation matrix under the maximizing the entropy criterion.
}
