% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_calculation_PCA.R
\name{S_calculation_PCA}
\alias{S_calculation_PCA}
\title{S-matrix computation (PCA)}
\usage{
S_calculation_PCA(Sigma, r, verbose = TRUE, tol = 0.001)
}
\arguments{
\item{Sigma}{Correlation matrix.}

\item{r}{The number of principal components conditioned on.}

\item{verbose}{Indicator of whether progress of computation is plotted.}

\item{tol}{Convergence tolerance}
}
\value{
A list of objects for knockoff generation.
}
\description{
This function computes the S-matrix from correlation matrix under the PCA knockoff framework.
}
