% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_calculation_SDP_no_Julia.R
\name{S_calculation_SDP_no_julia}
\alias{S_calculation_SDP_no_julia}
\title{S-matrix computation (SDP) without julia.}
\usage{
S_calculation_SDP_no_julia(Sigma, M, verbose = TRUE, tol = 0.001)
}
\arguments{
\item{Sigma}{Correlation matrix.}

\item{M}{The number of knockoff copies (under the multiple knockoff framework).}

\item{verbose}{Indicator of whether progress of computation is plotted.}

\item{tol}{Convergence tolerance}
}
\value{
Diagnoal elements of the S-matrix for knockoff generation.
}
\description{
This function computes the S-matrix from correlation matrix under the SDP criterion without julia.
}
