% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_calculation.R
\name{tau_calculation}
\alias{tau_calculation}
\title{Feature Statistics (Magnitude)}
\usage{
tau_calculation(x)
}
\arguments{
\item{x}{Feature importance scores of original feature and its knockoff copies.}
}
\value{
Feature statistics (magnitude)
}
\description{
This function computes feature statistics (magnitude) from importance scores of original feature and its knockoff copies.
}
\examples{
tau_calculation(1:6)
}
