% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_calculation.R
\name{Z_calculation}
\alias{Z_calculation}
\title{Z-scores computation}
\usage{
Z_calculation(p_value, log2FoldChange)
}
\arguments{
\item{p_value}{DGE p-values.}

\item{log2FoldChange}{Log2 fold change.}
}
\value{
DGE Z-scores.
}
\description{
This function computes Z-scores from DGE p-values (unadjusted) and log2 fold change.
}
